Simplify Startup Initializers
Since the project current modularization does not yet allow to split bindings and the rest of the code, moving the `Initializer`s registration to the `:app` module makes it easier to use.

- It avoids duplicated registrations (it was both present in `:app` and `:core:network`)
- It avoids removing registrations on other modules that would break during instrumented tests

`EnsureInitializerMetadata` had to be suppressed because the Lint rule does not actually check where the initializer are registered, and always expects to see it on the same module, which is quite a simplistic assumption.
https://googlesamples.github.io/android-custom-lint-rules/checks/EnsureInitializerMetadata.md.html

I also remove the hardcoded `androidx.startup` string with the proper string reference `@string/androidx_startup`.